requirements.txt
# For loading environment variables from .env
python-dotenv

# Google's library for the Gemini API
google-genai

# GUI
customtkinter